:elf: Use elf entity store for items
