+++ package.json
@@ -14,6 +14,9 @@
   "dependencies": {
     "@datorama/akita": "^6.2.4",
     "@headlessui/react": "^1.4.2",
+    "@ngneat/elf": "^1.0.2",
+    "@ngneat/elf-entities": "^2.0.2",
+    "@ngneat/react-rxjs": "^1.0.0",
     "chart.js": "^3.6.2",
     "dayjs": "^1.10.7",
     "jwt-decode": "^3.1.2",
