+++ shared/handlers/filter.handler.ts
@@ -1,10 +1,11 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/dist/client/router';
+import { useObservable } from '@ngneat/react-rxjs';
+import { selectAll } from '@ngneat/elf-entities';
+import { Item } from '../stores/items/item.model';
+import { getItemRarities, itemsStore } from '../stores/items/items.store';
 import { DropdownValue } from '../../components/styles/Dropdown';
 import useInventoryProvider from '../providers/inventory.provider';
-import useItemsProvider from '../providers/items.provider';
-import { Item } from '../stores/items/item.model';
-import { itemsQuery } from '../stores/items/items.query';
 import { Offer, Wish } from '../stores/markets/market.model';
 import { MARKET_TYPE } from './market.handler';
 
@@ -137,7 +138,7 @@ const FilterHandler = (type: FILTER_TYPE, itemsToLoad: number, trades?: Offer[] 
 
   const { inventory } = useInventoryProvider();
 
-  const { items } = useItemsProvider();
+  const [items] = useObservable(itemsStore.pipe(selectAll()));
 
   const [queryLoaded, setQueryLoaded] = useState(false);
   const [usernameChanged, setUsernameChanged] = useState(false);
@@ -293,8 +294,8 @@ const FilterHandler = (type: FILTER_TYPE, itemsToLoad: number, trades?: Offer[] 
             return 0;
           }
           if (itemFilterValues.orderBy.key === 'tagRarity') {
-            one = itemsQuery.getRarities().indexOf(one as string);
-            two = itemsQuery.getRarities().indexOf(two as string);
+            one = getItemRarities().indexOf(one as string);
+            two = getItemRarities().indexOf(two as string);
           }
           let returnValue = 0;
           if (one > two) {
@@ -325,8 +326,8 @@ const FilterHandler = (type: FILTER_TYPE, itemsToLoad: number, trades?: Offer[] 
           return 0;
         }
         if (itemFilterValues.orderBy.key === 'tagRarity') {
-          one = itemsQuery.getRarities().indexOf(one as string);
-          two = itemsQuery.getRarities().indexOf(two as string);
+          one = getItemRarities().indexOf(one as string);
+          two = getItemRarities().indexOf(two as string);
         }
         let returnValue = 0;
         if (one > two) {
