+++ shared/stores/items/items.service.ts
@@ -1,13 +1,12 @@
+import { setEntities } from '@ngneat/elf-entities';
 import { of } from 'rxjs';
 import { fromFetch } from 'rxjs/fetch';
 import { tap } from 'rxjs/operators';
 import { createNotification } from '../notification/notification.model';
 import { notificationService } from '../notification/notification.service';
-import { ItemsStore, itemsStore } from './items.store';
+import { itemsStore } from './items.store';
 
 export class ItemsService {
-
-  constructor(private itemsStore: ItemsStore) { }
 
   public fetchAllItems() {
     return fromFetch(`${process.env.NEXT_PUBLIC_BASE_API_URL}/api/items`).pipe(
@@ -16,7 +15,7 @@ export class ItemsService {
           next: async (res) => {
             const json = await res.json();
             if (res.ok) {
-              this.itemsStore.set(json);
+              itemsStore.update(setEntities(json));
             } else {
               const notification = createNotification({
                 content: json.message,
@@ -42,4 +41,4 @@ export class ItemsService {
 
 }
 
-export const itemsService = new ItemsService(itemsStore);
+export const itemsService = new ItemsService();
