+++ shared/stores/items/items.store.ts
@@ -1,18 +1,55 @@
-import { EntityState, EntityStore, StoreConfig } from '@datorama/akita';
+import { createState, Store } from '@ngneat/elf';
+import { withEntities } from '@ngneat/elf-entities';
 import { Item } from './item.model';
 
-export interface ItemsState extends EntityState<Item> { }
+const { state, config } = createState(withEntities<Item>());
 
-@StoreConfig({
-  name: 'items',
-  idKey: 'id'
-})
-export class ItemsStore extends EntityStore<ItemsState> {
+export const itemsStore = new Store({ name: 'items', state, config });
 
-  constructor() {
-    super();
-  }
+export function itemRarityToIngredientId(rarity: string): number {
+  const rarityToIngredientId: { [key: string]: number; } = {
+    common: 900,
+    uncommon: 901,
+    rare: 902,
+    veryrare: 903,
+    whimsical: 904
+  };
 
+  return rarityToIngredientId[rarity];
 }
 
-export const itemsStore = new ItemsStore();
+export function itemRarityToColor(rarity: string) {
+  switch (rarity) {
+    case 'common':
+      return '#7C7C7C';
+    case 'uncommon':
+      return '#ECECEC';
+    case 'unlock':
+      return '#C08051';
+    case 'eventrarity':
+      return '#FFE400';
+    case 'rare':
+      return '#45E53D';
+    case 'veryrare':
+      return '#1472FF';
+    case 'whimsical':
+      return '#FF00EA';
+    case 'promo':
+      return '#00E4FF';
+    default:
+      return '';
+  }
+}
+
+export function getItemRarities(): string[] {
+  return [
+    'common',
+    'uncommon',
+    'unlock',
+    'eventrarity',
+    'rare',
+    'veryrare',
+    'whimsical',
+    'promo'
+  ];
+}
